split files in preparation
